+++ .github/workflows/publish_pypi.yaml
@@ -17,13 +17,13 @@ jobs:
         with:
           python-version: '3.8' 
 
-    - name: Install build
-      run:  pip install build
+      - name: Install build
+        run:  pip install build
 
-    - name: Build a binary wheel and a source tarball
-      run:  python -m build --sdist --wheel --outdir dist/
+      - name: Build a binary wheel and a source tarball
+        run:  python -m build --sdist --wheel --outdir dist/
 
-    - name: Publish to PyPI
-      uses: pypa/gh-action-pypi-publish@master
-      with:
-        password: ${{ secrets.PYPI_API_TOKEN }}
+      - name: Publish to PyPI
+        uses: pypa/gh-action-pypi-publish@master
+        with:
+          password: ${{ secrets.PYPI_API_TOKEN }}
