+++ mabel/operators/__init__.py
@@ -3,3 +3,4 @@ from .end_operator import EndOperator
 from .filter_operator import FilterOperator
 from .noop_operator import NoOpOperator
 from .validator_operator import ValidatorOperator
+from .profile_data_operator import ProfileDataOperator
