+++ mabel/version.py
@@ -1,6 +1,6 @@
 # Store the version here so:
 # 1) we don't load dependencies by storing it in __init__.py
 # 2) we can import it in setup.py for the same reason
-__version__ = "0.4.1"
+__version__ = "0.4.2"
 
 #nodoc - don't add to the documentation wiki
